Fix inadvertently sealed traits.
