+++ src/base/mod.rs
@@ -1,10 +1,7 @@
 //! The base building blocks for working with ontology data.
-pub use term::simple::SimpleMinimalTerm;
-pub use term::AltTermIdAware;
-pub use term::{MinimalTerm, Term};
 pub use term_id::TermId;
 
-mod term;
+pub mod term;
 mod term_id;
 #[cfg(feature="pyo3")]
 pub(crate) mod py;
@@ -14,10 +11,10 @@ pub(crate) mod py;
 ///
 /// ## Examples
 ///
-/// [`crate::base::SimpleMinimalTerm`] implements `Identified`.
+/// [`crate::base::term::simple::SimpleMinimalTerm`] implements `Identified`.
 /// ```
 /// use ontolius::prelude::*;
-/// use ontolius::base::SimpleMinimalTerm;
+/// use ontolius::base::term::simple::SimpleMinimalTerm;
 ///
 /// let term_id = TermId::from(("HP", "1234567"));
 /// let term = SimpleMinimalTerm::new(term_id, "Sample term", vec![], false);
