+++ src/prelude.rs
@@ -1,8 +1,8 @@
 //! The list of recommended imports for using the library.
-pub use crate::base::AltTermIdAware;
+pub use crate::base::term::AltTermIdAware;
+pub use crate::base::term::MinimalTerm;
+pub use crate::base::term::Term;
 pub use crate::base::Identified;
-pub use crate::base::MinimalTerm;
-pub use crate::base::Term;
 pub use crate::base::TermId;
 
 pub use crate::error::OntographError;
